simple-cli-chat/requirements.txt
anthropic>=0.26.0
python-dotenv>=1.0.1
